Do not remove other unrelated collections

+++ src/crewai_expert/services/doc_files_chunking_service.py
@@ -189,6 +189,9 @@ class DocFilesChunkingService:
             return
 
         for collection in self._qdrant_client.get_collections().collections:
-            if collection.name != self._collection_name:
+            if (
+                collection.name != self._collection_name
+                and self._collection_name in collection.name
+            ):
                 self._qdrant_client.delete_collection(collection.name)
                 print(f"Deleted Qdrant collection '{collection.name}'")
